django-backend/requirements.txt
aenum==2.2.6
asgiref==3.11.1
bunq_sdk==1.28.0
certifi==2026.4.22
chardet==3.0.4
charset-normalizer==3.4.7
Django==6.0.4
django-cors-headers==4.9.0
Pillow>=11.1.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
idna==3.13
psycopg2-binary==2.9.12
pycryptodomex==3.23.0
PyJWT==2.12.1
requests==2.33.1
simplejson==3.20.2
sqlparse==0.5.5
urllib3==1.26.20